fix: [C-51] 🐞 Dark mode button in the navbar is corrected
