+++ frontend/app/src/common/layouts/core-layout/components/atoms/index.ts
@@ -1,0 +1,1 @@
+export * from './dark-mode/darkMode.component';
